rename leaf.rs -> primitive.rs
