+++ src/types.rs
@@ -1,7 +1,7 @@
 //! General wire types used in the compiler
 
 pub mod custom;
-pub mod leaf;
+mod primitive;
 mod signature;
 pub mod simple;
 pub mod type_enum;
